refactor: split domain schemas into entity modules
